requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.43.0
cron-descriptor==1.4.5
diff-match-patch==20241021
Django==4.2.20
django-celery-beat==2.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==24.3
django-import-export==4.3.7
django-jet-reboot==1.3.10
django-leaflet==0.31.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.9
exceptiongroup==1.2.2
freezegun==1.5.1
frozenlist==1.5.0
GDAL==3.10.2
geographiclib==2.0
geoip2==5.0.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
maxminddb==2.6.3
multidict==6.3.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
redis==5.3.0b5
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
